+++ lib/zen/components/wps_kit.rb
@@ -70,28 +70,52 @@ module Zen
         OpenStruct.new(file_hash: md5_digest, block_count:, file_stat: "")
       end
 
+      ##
+      # dotfiles , backup and restore USER_HOME dot config file
       def dotfiles
         # os home directory
         home = Dir.home
         archive_root = File.join(home, "Documents/Archive")
 
-        dotpath = "dotfiles"
+        dotpath = "dotfiles2"
 
         file_list = [
-          File.join(".zshrc"),
+          ".zshrc",
           ".zshenv",
           ".zprofile",
           ".profile",
           ".gitconfig",
           ".ssh",
-          ".m2/setting.xml"
+          ".m2/setting.xml",
+          ".rbenv/version",
+          ".pyenv/version"
         ]
 
         file_list.each do |file|
-          src_path = File.join(Dir.home, file)
-          out_path = File.join(archive_root, dotpath, file)
-          puts "dotfiles:#{src_path}, #{out_path}"
+          src_file = File.join(Dir.home, file)
+          to_file = File.join(archive_root, dotpath, file)
+
+          to_dir = File.dirname(to_file)
+          puts "dotfiles:#{src_file}, #{to_file}"
+
+          FileUtils.makedirs(to_dir) unless File.exist?(to_dir)
+
+          next unless File.exist?(src_file)
+
+          # system("cp -fr #{src_file}  #{to_dir}")
+          FileUtils.cp_r(src_file, to_dir)
+          puts "done"
         end
+      end
+
+      ##
+      # Use tar  and zstd compression and decompression directory
+      def zstds(from_dir, output_filename)
+        # Use zstd if desired, otherwise comment out the line
+        system("tar -cf - #{from_dir} | zstd -f > #{output_filename}")
+
+        # Alternative for gzip compression (replace zstd with gzip)
+        # system("tar -cf - #{_from} | gzip > #{output_filename}")
       end
     end
   end