+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
+## v0.1.1
+* `wps dotfiles`, backup and restore dotfiles
+* `wps zstds`, use tar and zstd to compress and decompress directory
+
 ## v0.1.0
 * starter feature ,initial project, add feature to project.
-* starter workspace 
-* wps fdupes ,find your workspace duplicates file
+* `starter workspace` ,initial workspace directory
+* `wps fdupes` ,find your workspace duplicates file